Finishes the web server!
Can now view and browse journal entries, and can even search through
them, click and be directed there (under the corresponding heading of
course).

Took me a while to perfect, but every search request take around 1
millisecond. This is an O(n) search algorithm, but with my database
of 124241 characters of text, it is good enough at the moment.